Remove CefGpuProcess - wasn't used for anything Simplify the class structure for the render/gpu subprocesses
