--- CefSharp.BrowserSubprocess/CefGpuProcess.cs
@@ -1,16 +1,0 @@
-﻿// Copyright © 2010-2015 The CefSharp Authors. All rights reserved.
-//
-// Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
-
-using System.Collections.Generic;
-
-namespace CefSharp.BrowserSubprocess
-{
-    public class CefGpuProcess : CefSubProcess
-    {
-        public CefGpuProcess(IEnumerable<string> args) 
-            : base(args)
-        {
-        }
-    }
-}
